refactor: Логика группирования данных офферов вынесена в отдельную функцию
